+++ src/game/physics/Physics_Base.h
@@ -70,9 +70,9 @@ public: // common physics interface
 	const idBounds &GetBounds(int id = -1) const;
 	const idBounds &GetAbsBounds(int id = -1) const;
 
-	bool Evaluate(int timeStepMSec, int endTimeMSec);
-	void UpdateTime(int endTimeMSec);
-	int GetTime(void) const;
+	bool Evaluate(float timeStepMSec, float endTimeMSec);
+	void UpdateTime(float endTimeMSec);
+	float GetTime(void) const;
 
 	void GetImpactInfo(const int id, const idVec3 &point, impactInfo_t *info) const;
 	void ApplyImpulse(const int id, const idVec3 &point, const idVec3 &impulse);
